refactor: creates opt module
With optimization related code
